requirements.txt
psycopg2-binary
fastapi
"uvicorn[standard]"
PyGithub
openai
langchain